pyproject.toml
[project]
name = "supermew"
version = "0.1.0"
description = "SuperMew - A cute cat bot with RAG capabilities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "rich>=14.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.8.0",
    "langchain>=0.2.14",
    "langchain-core>=0.2.37",
    "langchain-community>=0.2.12",
    "langchain-text-splitters>=0.2.2",
    "langchain-openai>=0.1.22",
    "langchain-chroma>=0.1.0",
    "langchain-huggingface>=0.1.0",
    "langgraph>=0.2.31",
    "pypdf>=4.3.1",
    "docx2txt>=0.8",
    "chromadb>=0.5.5",
    "protobuf<=3.20.3",
    "sentence-transformers>=2.7.0",
]

[project.optional-dependencies]
study = [
    "langchain-classic>=0.2.0",
    "bilibili-api-python>=17.0.0",
]
